more work on edit logic

+++ src/utils/DBManager.java
@@ -140,6 +140,16 @@ public class DBManager {
     	int randNum = rand.nextInt((questions.size()-1) + 1);
     	return questions.get(randNum);
     }
+
+	public void updateQuestion(Question question){
+		em.getTransaction().begin();
+		for(Answer answer: question.getAnswers()){
+			em.persist(answer);
+		}
+		em.merge(question);
+
+		em.getTransaction().commit();
+	}
     
     public int commitQuestion(Question question){
     	
@@ -148,6 +158,7 @@ public class DBManager {
     		em.persist(answer);
     	}
     	em.persist(question);
+
     	em.getTransaction().commit();
     	
     	return question.getId();
@@ -302,12 +313,27 @@ public class DBManager {
 
 	public void removeQuestionById(int id){
 		Question question = getQuestionById(id);
+		if(question != null){
+			em.getTransaction().begin();
 
-		em.getTransaction().begin();
+			em.remove(question);
 
-		em.remove(question);
+			em.getTransaction().commit();
+		}
 
-		em.getTransaction().commit();
+	}
+
+	public void removeAnswerById(int id){
+		Answer answer = getAnswerById(id);
+
+		if(answer != null){
+			em.getTransaction().begin();
+
+			em.remove(answer);
+
+			em.getTransaction().commit();
+		}
+
 	}
 
     public void close(){
